improve Controller Components
