+++ frontend/src/javascripts/poject/components/Controller/Styles.css
@@ -1,0 +1,8 @@
+.ControllerButtomWrap {
+  width: 220px;
+  height: 35px;
+}
+.Clear {
+  clear: both;
+  height: 10px;
+}
