+++ frontend/src/javascripts/poject/components/Field/Styles.css
@@ -10,11 +10,3 @@
   height: 300px;
   width: 260px;
 }
-.Clear {
-  clear: both;
-  height: 10px;
-}
-.ControllerButtomWrap {
-  width: 220px;
-  height: 35px;
-}
